Criando signal de cadastro de usuário
