+++ templates/email/email_welcome.html
@@ -1,0 +1,3 @@
+Boas Vindas!
+Seja bem vindo ao nosso sistema! Gostaríamos de agradecer por se registrar, entre
+em contato com o administrador do sistema solicitando a liberação do seu cadastro! Equipe You Service!
